Begin component reorganization
- Move core UI components to their new locations:
  - Buttons: button moved to ui/buttons
  - Inputs: input moved to ui/inputs
  - Layout: card moved to ui/layout
  - Feedback: toast and toaster moved to ui/feedback
- Move layout components to layout directory:
  - Header
  - Footer
- Create index files for all new directories
- Update import paths in moved components
- Maintain backward compatibility by keeping original files as proxies with deprecation notices
