+++ src/components/layout/Header.tsx
@@ -1,13 +1,7 @@
-
-/**
- * @deprecated This file is maintained for backward compatibility. 
- * Import from '@/components/layout/Header' instead.
- */
-
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { Menu, X } from 'lucide-react';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/buttons/button';
 import { menuItems } from '@/data/menuItems';
 import DesktopNavigation from '@/components/navigation/DesktopNavigation';
 import MobileMenu from '@/components/navigation/MobileMenu';
